refactor the Body Footer
